+++ packages/database/src/test/example/entities/extended-outbox.entity.ts
@@ -1,0 +1,5 @@
+import { OutboxEntity } from '../../../outbox/entities/outbox.entity.js';
+
+export class ExtendedOutboxEntity extends OutboxEntity {
+  channel: string = 'default';
+}