+++ packages/database/src/test/example/models/extended-outbox.model.ts
@@ -1,0 +1,8 @@
+import { Column, Entity } from 'typeorm';
+import { OutboxModel } from '../../../outbox/models/outbox.model.js';
+
+@Entity('outbox_extended')
+export class ExtendedOutboxModel extends OutboxModel {
+  @Column({ type: 'varchar', length: 50, default: 'default' })
+  channel: string = 'default';
+}